refactor: ControllerDto를 ServiceDto로 변환하는 메소드 추가
