+++ src/main/java/roomescape/controller/dto/ReservationSaveRequestDto.java
@@ -1,0 +1,13 @@
+package roomescape.controller.dto;
+
+import roomescape.service.dto.ReservationSaveServiceDto;
+
+public class ReservationSaveRequestDto {
+    private String name;
+    private String date;
+    private String time;
+
+    public ReservationSaveServiceDto toServiceDto() {
+        return new ReservationSaveServiceDto(name, date, time);
+    }
+}
