--- src/main/java/roomescape/domain/dto/ReservationSaveRequestDto.java
@@ -1,20 +1,0 @@
-package roomescape.domain.dto;
-
-public class ReservationSaveRequestDto {
-    private String name;
-    private String date;
-    private String time;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-}
