+++ src/main/java/roomescape/service/RoomescapeService.java
@@ -2,11 +2,11 @@ package roomescape.service;
 
 import java.util.List;
 import roomescape.domain.Reservation;
-import roomescape.domain.dto.ReservationSaveRequestDto;
+import roomescape.service.dto.ReservationSaveServiceDto;
 
 public interface RoomescapeService {
     List<Reservation> getReservations();
-    Reservation save(ReservationSaveRequestDto reservation);
+    Reservation save(ReservationSaveServiceDto reservation);
     boolean deleteById(long id);
 
 }
