+++ src/main/java/roomescape/service/dto/ReservationSaveServiceDto.java
@@ -1,0 +1,25 @@
+package roomescape.service.dto;
+
+public class ReservationSaveServiceDto {
+    private final String name;
+    private final String date;
+    private final String time;
+
+    public ReservationSaveServiceDto(String name, String date, String time) {
+        this.name = name;
+        this.date = date;
+        this.time = time;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public String getTime() {
+        return time;
+    }
+}
